requirements.txt
fastapi[standard]
uvicorn
boto3
requests
requests-aws4auth
python-dotenv
pymupdf